refactor caluclator divide function to use double types, not int

+++ src/main/java/Calculator.java
@@ -21,7 +21,7 @@ public class Calculator {
         return valueOne * valueTwo;
     }
 
-    public int divide(int valueOne, int valueTwo){
+    public double divide(double valueOne, double valueTwo){
         return valueOne / valueTwo;
     }
 }
